+++ extend/middleware/logger.go
@@ -11,51 +11,6 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// 日志记录到文件
-func LoggerToFile() gin.HandlerFunc {
-	logger := app.GetLogger()
-
-	return func(c *gin.Context) {
-		// 开始时间
-		startTime := time.Now()
-
-		// 处理请求
-		c.Next()
-
-		// 结束时间
-		endTime := time.Now()
-
-		// 执行时间
-		latencyTime := endTime.Sub(startTime)
-
-		// 请求方式
-		reqMethod := c.Request.Method
-
-		// 请求路由
-		reqUri := c.Request.RequestURI
-
-		// 状态码
-		statusCode := c.Writer.Status()
-
-		// 请求IP
-		clientIP := c.ClientIP()
-
-		// 用户代理
-		userAgent := c.Request.UserAgent()
-
-		// 日志格式
-		logger.WithFields(map[string]any{
-			"request_id":   ctxkeys.GetRequestID(c),
-			"status_code":  statusCode,
-			"latency_time": latencyTime,
-			"client_ip":    clientIP,
-			"req_method":   reqMethod,
-			"req_uri":      reqUri,
-			"user_agent":   userAgent,
-		}).Info("HTTP Request")
-	}
-}
-
 func RequestID() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		rid := c.GetHeader("X-Request-ID")